+++ src/index.js
@@ -3,15 +3,20 @@ const OGR = require('./lib/ogr')
 const random = require('randomstring')
 const rimraf = require('rimraf')
 const _ = require('highland')
+const mkdirp = require('mkdirp')
 
 function createStream (format, options) {
+  options = options || {}
   options.path = `${options.path || '.'}/${random.generate()}`
+  mkdirp.sync(options.path)
   const stream = _.pipeline(stream => {
+    const ogrStream = OGR.createStream(format, options)
     return stream
     .pipe(VRT.createStream(options))
     .on('log', l => stream.emit('log', l))
     .on('error', e => stream.emit('error', e))
-    .pipe(OGR.createStream(format, options))
+    .on('properties', p => ogrStream.emit('properties', p))
+    .pipe(ogrStream)
     .on('log', l => stream.emit('log', l))
     .on('error', e => stream.emit('error', e))
     .on('end', e => rimraf(options.path, err => {